+++ Round.js
@@ -1,0 +1,11 @@
+
+class Round {
+    /**
+     * Creates an instance of Round with an empty list of matches.
+     */
+        constructor() {
+            this.match = [];
+        }
+    }
+
+    export default Round;
